Add files via upload
The Supplemental material for MLFD algorithm. The experimental results of MLFD algorithm and other nine comparison algorithms when dimension D = 10, 30 and 50 were presented in the supplementary material. In addition, for these three dimensions, the statistical comparisons of the WSRT for MLFD algorithm versus other algorithms were also shown in the supplementary materials. Meanwhile, the average ranking of Friedman test for MLFD and other comparison algorithms were also shown in the supplementary materials. Finally, the convergence curves of 10 algorithms under 17 test functions in these three dimensions will also be shown in supplementary materials.